+++ inselect/gui/cookie_cutter_choice.py
@@ -80,6 +80,14 @@ class CookieCutterChoice(QObject):
         self.load(path)
 
     @property
+    def current_path(self):
+        """The path to the selected UserTemplate or None, if the default
+        template is selected
+        """
+        current = QSettings().value(self.PATH_KEY)
+        return Path(current) if current else None
+
+    @property
     def current(self):
         "The selected CookieCutter"
         return self._current