+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This is an overview of major changes. Refer to the git repository for a full log
 Version 0.1.34
 -------------
 - #325 Command-line option to set window size
+- #328 Menu commands to show cookie cutter and metadata template files
 
 Version 0.1.33
 -------------